feature: supplier update has been made and the branch has been cleaned up as well
